% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msePT.R
\name{FUNCTION}
\alias{FUNCTION}
\alias{msePT}
\title{An example function to carry out an MSE run for a given MP}
\usage{
msePT(omp, sr, cpue, cpuesel, years, hcr = ~ifelse(dep <= Dlimit, 0,
  ifelse(dep < Dtarget, (lambda * MSY)/(Dtarget - Dlimit) * (dep - Dlimit),
  lambda * MSY)), hcrparams = FLPar(Dlimit = 0.1, Dtarget = 0.4, lambda = 1,
  dltac = 0.15, dhtac = 0.15), dlag = 1, mlag = 1, oemparams = FLPar(sd =
  0, b = 0), imparams, tune = FALSE, verbose = FALSE, sa = TRUE)
}
\arguments{
\item{PARAM}{Lorem ipsum dolor sit amet}
}
\value{
RETURN Lorem ipsum dolor sit amet
}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. 
Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id 
nunc euismod ante fringilla lobortis. Aliquam ullamcorper in diam non placerat.
}
\details{
Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa, 
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in 
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae 
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien, 
feugiat blandit arcu in, dapibus dictum felis.
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{design}
