% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mseIndex.R
\name{mseIndex}
\alias{mseIndex}
\title{mseIndex}
\usage{
mseIndex(omp, sr, cpue, cpuesel, years, verbose = FALSE, hcr = ~tac * (1 +
  lambda * slope), hcrparams = FLPar(lambda = 1.25, ny = 5, dltac = 0.15,
  dhtac = 0.15), dlag = 1, mlag = 1, oemparams = FLPar(sd = 0, b = 0),
  imparams, tune = FALSE)
}
\arguments{
\item{omp}{PARAM_DESCRIPTION}

\item{sr}{PARAM_DESCRIPTION}

\item{cpue}{PARAM_DESCRIPTION}

\item{cpuesel}{PARAM_DESCRIPTION}

\item{years}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{hcr}{PARAM_DESCRIPTION, Default: ~tac * (1 + lambda * slope)}

\item{hcrparams}{PARAM_DESCRIPTION, Default: FLPar(lambda = 1.25, ny = 5, dltac = 0.15, dhtac = 0.15)}

\item{dlag}{PARAM_DESCRIPTION, Default: 1}

\item{mlag}{PARAM_DESCRIPTION, Default: 1}

\item{oemparams}{PARAM_DESCRIPTION, Default: FLPar(sd = 0, b = 0)}

\item{imparams}{PARAM_DESCRIPTION}

\item{tune}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
data(cod)
}
\seealso{
\code{\link[utils]{txtProgressBar}},\code{\link[utils]{setTxtProgressBar}}
}
